update tool type to be partial in store for editing
