+++ app/src/pages/playground/PlaygroundTools.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 
-import { Button, Card, Flex, Icon, Icons } from "@arizeai/components";
+import { Card, Flex } from "@arizeai/components";
 
-import { JSONToolEditor } from "@phoenix/components/code";
 import { usePlaygroundContext } from "@phoenix/contexts/PlaygroundContext";
 
+import { PlaygroundTool } from "./PlaygroundTool";
 import { TitleWithAlphabeticIndex } from "./TitleWithAlphabeticIndex";
 import { PlaygroundInstanceProps } from "./types";
 
@@ -20,7 +20,6 @@ export function PlaygroundTools(props: PlaygroundToolsProps) {
   if (instance == null) {
     throw new Error(`Playground instance ${instanceId} not found`);
   }
-  const updateInstance = usePlaygroundContext((state) => state.updateInstance);
   if (instance.tools == null) {
     throw new Error(`Playground instance ${instanceId} does not have tools`);
   }
@@ -35,50 +34,16 @@ export function PlaygroundTools(props: PlaygroundToolsProps) {
       variant="compact"
     >
       <Flex direction="column" gap="size-50">
-        {instance.tools.map((tool) => (
-          <Card
-            collapsible
-            variant="compact"
-            key={tool.id}
-            title={tool.definition.function.name}
-            bodyStyle={{ padding: 0 }}
-            extra={
-              <Button
-                aria-label="Delete tool"
-                icon={<Icon svg={<Icons.TrashOutline />} />}
-                variant="default"
-                size="compact"
-                onClick={() => {
-                  updateInstance({
-                    instanceId,
-                    patch: {
-                      tools: instance.tools.filter((t) => t.id !== tool.id),
-                    },
-                  });
-                }}
-              />
-            }
-          >
-            <JSONToolEditor
-              value={JSON.stringify(tool.definition, null, 2)}
-              onChange={(value) => {
-                updateInstance({
-                  instanceId,
-                  patch: {
-                    tools: instance.tools.map((t) =>
-                      t.id === tool.id
-                        ? {
-                            ...t,
-                            definition: JSON.parse(value),
-                          }
-                        : t
-                    ),
-                  },
-                });
-              }}
+        {instance.tools.map((tool) => {
+          return (
+            <PlaygroundTool
+              key={tool.id}
+              playgroundInstanceId={instanceId}
+              tool={tool}
+              instanceTools={instance.tools}
             />
-          </Card>
-        ))}
+          );
+        })}
       </Flex>
     </Card>
   );
