+++ app/src/store/playground/playgroundStore.tsx
@@ -113,9 +113,6 @@ export function createTool(toolNumber: number): Tool {
               type: "string",
             },
           },
-          required: [],
-          additionalProperties: false,
-          strict: false,
         },
       },
     },
